# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/addons/:name/tasks": {
            "post": {
                "description": "Create an addon task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create an addon task.",
                "parameters": [
                    {
                        "description": "Task data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                }
            }
        },
...
        "/tasks": {
            "get": {
                "description": "List all tasks.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "List all tasks.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create a task.",
                "parameters": [
                    {
                        "description": "Task data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Get a task by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get a task by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a task.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update a task.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete a task.",
                "tags": [
                    "delete"
                ],
                "summary": "Delete a task.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/tasks/{id}/report": {
            "put": {
                "description": "Update a task report.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update a task report.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TaskReport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TaskReport data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TaskReport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TaskReport"
                        }
                    }
                }
            },
            "post": {
                "description": "Update a task report.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Create a task report.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TaskReport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TaskReport data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TaskReport"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.TaskReport"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.Task": {
            "type": "object",
            "required": [
                "addon",
                "name"
            ],
            "properties": {
                "addon": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "isolated": {
                    "type": "boolean"
                },
                "job": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "report": {
                    "$ref": "#/definitions/api.TaskReport"
                },
                "started": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "terminated": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "api.TaskReport": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string"
                },
                "completed": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "task": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
...
    }
}